requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.6
google-genai
PyMuPDF==1.23.8
celery==5.3.6
redis==5.0.1
alembic==1.13.1
python-dotenv==1.0.0
beautifulsoup4==4.12.3
python-docx==1.1.2
